Scene Collection Traversal experimenting.
